funciones en carpeta lib
