+++ src/lib/firebase.js
@@ -9,13 +9,13 @@
  // Your web app's Firebase configuration
  // For Firebase JS SDK v7.20.0 and later, measurementId is optional
  const firebaseConfig = {
-   apiKey: "AIzaSyDOU7BlIf_BBE7_MUDfkNFlAuvCWqscfMg",
-   authDomain: "social-network-17eb4.firebaseapp.com",
-   projectId: "social-network-17eb4",
-   storageBucket: "social-network-17eb4.appspot.com",
-   messagingSenderId: "775373625184",
-   appId: "1:775373625184:web:ac1209b244e2d83bdcf005",
-   measurementId: "G-7M4HS38DST"
+  apiKey: "AIzaSyDOU7BlIf_BBE7_MUDfkNFlAuvCWqscfMg",
+  authDomain: "social-network-17eb4.firebaseapp.com",
+  projectId: "social-network-17eb4",
+  storageBucket: "social-network-17eb4.appspot.com",
+  messagingSenderId: "775373625184",
+  appId: "1:775373625184:web:ac1209b244e2d83bdcf005",
+  measurementId: "G-7M4HS38DST"
  };
 
  // Initialize Firebase
